requirements.txt
fastapi
uvicorn
Pillow
tensorflow==2.13.0
mediapipe
python-multipart
